+++ snovault/server_defaults.py
@@ -21,7 +21,6 @@ exported(
 )
 
 
-
 ACCESSION_FACTORY = __name__ + ':accession_factory'
 ACCESSION_PREFIX = app_project().ACCESSION_PREFIX
 ACCESSION_TEST_PREFIX = 'TST'